scripts/contrib: Add override conversion script
This adds a script developed to migrate metadata to use the new override
syntax. It is a bit rough but since its for a single use with validation, it
doesn't need to be perfect. It is run simply as:

scripts/contrib/convert-overrides.py <directory>

For other layers it would need updating for override names and exclusions for
functions/variable names with "append", "prepend" or "remove" in them.

(From Yocto kirkstone-4.0.5 rev: fbdf93f43ff4b876487e1f26752598ec8abcb46e)

Signed-off-by: Anton Mikanovich <amikan@ilbers.de>