expand-on-first-boot: switch back away from systemd-growfs
systemd-growfs in some versions has runtime deps on symlinks created
by udev and we have a race there that shows especially on smaller/slower
devices.
We had patches dealing with that but they ended up looking rather
complex.
The gain of systemd-growfs is potentially a bunch more filesystems. But
we can add them as well in sort of a switch-case. We never used it the
way it is intended because it did not really fit, so better not use it.

Signed-off-by: Henning Schild <henning.schild@siemens.com>