The expand last partition rework and generalisation
The script has been reowrked to simplify its code:

 - the last partition free space is not computed anymore because useless
 - on exit a function take care of some final tasks and logs in /dev/kmsg

and to extend its functionalities supporting new filesystems:

 - filesystem type is managed by a switch-case block of code
 - temporary directory to mount is created relying on TMPDIR
 - unsupervised systems have the option to reboot because error
 - if temporary directory creation fails 2nd try with /dev/shm
 - additional ext2, ext3 and btrfs filesystems supported
 - udev support: udevadm settle is the first task to run
 - unpartitioned volumes are resized as well (tested)

This patch has been created by a collection of 7 patches to be applied on
ilbers ISAR, branch next, commit d26660b724b034b602f3889f55a23cd9be2e87bd

   Author: Roberto A. Foglietta <roberto.foglietta@gmail.com>

This code has been developed taking inspiration by suggestions and code from:

   Felix Moessbauer <felix.moessbauer@siemens.com>
   Tobias Schmidl <tobiasschmidl@siemens.com>
   Henning Schild <henning.schild@siemens.com>

Signed-off-by: Roberto A. Foglietta <roberto.foglietta@gmail.com>

= 1/7 == In expand last partition wait for udev have finished =================

bugfix, expand last partition, wait for udev completion

On some hardware the udev did not create the links to the disk partitions
while this script is working thus it fails. This patch fixes those cases.

v2: as underlined by Tobias this script requires systemd-udevd.service and
it can safely run only after that service has been started otherwise it might
fail. This dependency will reorder the boot procedure but not slow down it
unless the hardware is slow to get mapped by udev but in that case resize
the last partition would have failed. Thus, no regression is expected.

= 2/7 == In expand last partition script btrfs support added ==================

Improvement, expand-on-first-boot: support for btrfs added

v2: the mount point for btrfs filesystem to resize is under /dev/shm
by default but a variable defined in the running environment can
change the temporary directory path. The /dev/shm has been chosen
because in some systems /tmp can be on a RO root filesystem and
/tmp not yet mounted with tmpfs. This requires the system to have
/dev/shm configured into the kernel.

v3: the previous version was using a bashism but the shebang is /bin/sh
this means that the script will fail in a customized system in which /bin/sh
do not link to bash or bash is not available.

v4: Felix suggested using mktemp for the temporary directory in which to
mount the btrfs filesystem to resize. The default path /tmp could be changed
using the variable TMPDIR to set in the running environment.

v5: typo fixes into the patch description

v6: mktemp does a reasonable second try with -p /dev/shm before failing

= 3/7 == If expand last partition fails, unsupervised systems reboot ==========

Improvement for unsupervised embedded devices which need to reboot if expanding
last partition fails because an under-provided system should not even complete
the boot in the field in order to avoid functioning failures at unexpected
future time. This patch do not affect the previous behaviour because requires
that /etc/unsupervised file exists.

v2: if there is no partition table exit 0 instead of exit 1

= 4/7 == In expand last partition checking the size is useless ================

code maintenance, superfluous operations commented, kept for the future

Tests shown that resize in full a ext4 or btrfs partition/filesystem does not
hurt and does not report a failure, so nothing changes without the size check

= 5/7 == The expand last partition might support full volumes also, to test ==

Improvement, expansion of unpartitioned volumes supported also but to test

DONE, 14.12.2022: this change should be tested with a single volume system
preliminary tests shown that ext4 and btrfs entire volumes cam be resized

= 6/7 == In expand last partition trap exitnlog EXIT logs with simpler code ===

Improvement, the script log at exit with the status of completion and its full
path in this way - in case of failure - the user will find in dmesg the error
and also the script full path to debug. It uses /dev/kmsg to log.
Code simplified.

= 7/7 == In expand last partition few improvements and a bugfix ===============

Some improvements and a bugfix in expanding last partition.

Improvements: filesystems ext2 and ext3 also supported, warning in dmesg in
case of unsupported filesystem, the variable tmpdir is used with double
quotes even if unneeded but be consistent with the original script notation.

Bugfix: even if btrfs resize fails, umount the last partition and return error.