In expand last partition few improvements and a bugfix
Some improvements and a bugfix in expanding last partition.

Improvements: filesystems ext2 and ext3 also supported, warning in dmesg in
case of unsupported filesystem, the variable tmpdir is used with double
quotes even if unneeded but be consistent with the original script notation.

Bugfix: even if btrfs resize fails, umount the last partition and return error.

Mention: Henning

Signed-off-by: Roberto A. Foglietta <roberto.foglietta@gmail.com>