In expand last partition recipe udev pacakge added
A dependency needed added, this grants that udevd and udevadm will be installed

Many systems might not need udevadm settle for expanding last partition but few
others which rely on separated couple of bootfs/rootfs where last partition is on
a volume connected through a kind of bus which not be fully ready at boot time
until udevd will end to populate the /dev tree.

Moreover, expanding the last partition is a single time task which runs only on
the first machine boot and then will be disabled forever. Thus, even if it
would take some time on some systems, it is acceptable in order to grant to the
widest variety of systems this service. Alternative is to accept that some
system will be left unprovided of the storage space they have been designed for.

Signed-off-by: Roberto A. Foglietta <roberto.foglietta@gmail.com>