scripts/ci_build.sh installs dependencies and uses mktemp
CI build requires some commands in order to operate properly and some of these
are checked to be in the executable path but nothing more. This fix the issue
installing them at the beginning of the script.

Previous user or concurrent users may use scripts/ci_build.sh and in
particular root user creates a temporary directory that is not removable by
common users. Using a combination of mktemp and trap on EXIT will solve
all of these problems.

Signed-off-by: Roberto A. Foglietta <roberto.foglietta@gmail.com>