changes for a faster build using less disk space, p.1
Improvement, faster build for larger projects with a lot of debian packages

This patch leverages few ways to accomplish the task to optimise the large
buildings. It uses hard physical links instead of copying files which is a
faster way and saves disk spaces. Put a CACHEDIR.TAG file in each cache folder
in such a way every tar that uses the option --exclude-caches could be avoided
to include debian packages saving a lot of time and disk space.

This functions were used to make a comparison about the improvement:

	function fds() {
	  df -m / | grep /dev | tr -s ' ' | cut -d' ' -f4;
	}

	function ms() {
	  du -ms "$@" 2>/dev/null | cut -f1
	}

	function chkfds() {
	  st=$(fds); mx=0;
	  while sleep 1; do ct=$[st-$(fds)];
	  if [ $ct -gt $mx ]; then printf "%5s Mb\n" $ct; mx=$ct; fi;
	  done; ct=$[st-$(fds)];
	  printf "%5s Mb (max)\n" $mx;
	  printf "%5s Mb (rest)\n" $ct;
	  printf "%5s Mb (deb)\n" $(ms build/downloads/deb);
	  printf "%5s Mb (wic)\n" $(ms build/tmp/deploy/images/*/*.wic);
	  printf "%5s Mb (cache)\n" $(ms build/sstate-cache);
	}

	chkfds & read; killall sleep; sleep 1; echo

on another terminal:

	clean all
	build basic-os

when complete press a key in the previous terminal to collect data:

	results before and after the changes

	43954 Mb (max)   |  8657 Mb (max)
	26548 Mb (rest)  |  4118 Mb (rest)
	 3741 Mb (deb)   |  3741 Mb (deb)
	  820 Mb (wic)   |   820 Mb (wic)
	11789 Mb (cache) |   579 Mb (cache)
	time: 8m33s      | time: 4m34s

After the changes the max disk size used is 2x downloaded debian packages
plus the size of the wic image and the debootstrap cache:

	2 x 3741 + 820 + 579 = 8881 > 8654 MB

About the building time, it has been reduced by 2x times.

Signed-off-by: Roberto A. Foglietta <roberto.foglietta@gmail.com>