deb_dl_dir_im/export rework, also transfer the repositories lists v2
Most of the times, not only the .deb packages but also the repositories
lists should be transfered to improve performances. In fact, without
this lists, apt-get update will download again and again.

This patch is the result of a squash of other two previous patches.

v.2: includes the removing of a line which removes lists after they
     have been imported neutralising the benefit of v.1 patch

PATCH 1: deb_dl_dir_im/export nolists when USE_CCACHE is not active

The deb-dl-dir class has been partially reworked in order to transfer
more efficiently the packages (ln -P instead of cp -n) and to transfer
also the data useful for apt-get update in such a way it will not repeat
the same download for each recipe but just the diff, if nay.

* 0f95199 - deb-dl-dir deb_dl/lists_dir_im/export rationalisation, p.3
* 59aa192 - deb-dl-dir deb_dl/lists_dir_im/export rationalisation, p.2
* 051d238 - deb-dl-dir deb_dl/lists_dir_im/export rationalisation
* 310417a - deb-dl-dir deb_lists_dir_im/export added + rationalisation

PATCH 2: deb-dl-dir: deb_dl_dir_import/export bugfixed and code reworked

Signed-off-by: Roberto A. Foglietta <roberto.foglietta@gmail.com>