[RAF] License changed in GPLv3 for the composition, check the README.md
Almost all the files are under MIT license or GPLv2/2+ and the others are in
the public domain due to their simplicity and/or standardisation like system
configuration. However the composition of these files is protected by the GPLv3
license (since January 4th, 2023).

This means that everyone can use a single MIT licensed file or a part of it
under the MIT license terms. Instead, using two of them or two parts of them
implies that you are using a subset of this collection. Thus a derived work of
this collection which is licensed under the GPLv3 also.

The GPLv3 license applies to the composition unless you are the original
copyright owner or the author of a specific unmodified file. This means that
every one that can legally claim rights about the original files maintains its
rights, obviously. So, it should not need to complain with the GPLv3 license
applied to the composition. Unless, the composition is adopted for the part
which had not the rights, before.

Some execption could apply to some specific version of few files, please
check always the header to be sure.

v.2: this commit is based on this original submited change

    License changed in GPLv3 for the composition, check the README.md
    Author: Roberto A. Foglietta <roberto.foglietta@gmail.com>
    Date: Wed Jan 4 12:53:20 2023 +0100

Signed-off-by: Roberto A. Foglietta <roberto.foglietta@gmail.com>