imagetypes wic: generate_wic_image() artifacts removal on err exit
imagetypes_wic, generate_wic_image(): improvements / bugfix

When the user stops with CTRL-C the building and the generate_wic
_image() is interrupted, then partial artifacts remain on the
disk and the user could be confused by them and induced in error.
Also the custom/QA-testing scripts can be induced in error. Thus
removing these uncompleted artifacts is necessary and a bugfix in
general terms.

Signed-off-by: Roberto A. Foglietta <roberto.foglietta@gmail.com>