[RAF] perf. suite script and presentation of the results v8
THe performance suite scripts and presentation of the results

v.2: some blanks spaces removed in doc/build-optim.md

v.3: new script: check-build-perform.sh [build_target]

v.4: new script: isar-next-rebase.sh [-r]

v.5: isar-next-rebase.sh rpull the local branch before rebasing

v.6: script update: check-build-perform.sh

v.7: performances showcase added in the commit description

v.8: isar-next-rebase.sh updated

PERFORMANCES SHOWCASES
======================

Compared to the original the performances increase is impressive, in both cases:

          original          roban74:evo2
       ------------ basic-os -------------- ==============
       43954 Mb (max)   |  3498 Mb (max)    12.57x  16.07x
       26548 Mb (rest)  |  2588 Mb (rest)   10.26x  14.50x
        3741 Mb (deb)   |  3417 Mb (deb)      -
         820 Mb (wic)   |   814 Mb (wic)      -
       11789 Mb (cache) |   273 Mb (cache)  43.18x
       time: 8m33s      | time: 3m09s        2.71x

          original          roban74:evo2
       ------------ complete ------------- ===============
       52507 Mb (max)   | 28606 Mb (max)     1.84x   2.23x
       43311 Mb (rest)  | 19413 Mb (rest)    2.23x   3.33x
        3741 Mb (deb)   |  3417 Mb (deb)      -
        9159 Mb (wic)   |  9155 Mb (wic)      -
       11799 Mb (cache) |   283 Mb (cache)  41.69x
       time: 20m13s     | time: 11m44s       1.72x

Considering that the local .deb cache growing with the time because new updates
will be added, and comparing the performance gains between the basic-os and the
complete image, we can say that on the long run also the complete case will
reach the 2x of performance at least in time building.

Running the performances test in a VirtualBox with 2 threads CPU and 2GB of RAM:

          vm:next             vm:evo2
       ------------ basic-os -------------- ==============
       30947 Mb (max)   |  3537 Mb (max)     8.75x  11.09x
       24384 Mb (rest)  |  2559 Mb (rest)    9.53x  13.55x
        3496 Mb (deb)   |  3496 Mb (deb)      -
         814 Mb (wic)   |   820 Mb (wic)      -
       10752 Mb (cache) |   233 Mb (cache)  46.15x
       time: 12m34s     | time: 3m09s        3.99x

          vm:next             vm:evo2
       ------------ complete ------------- ===============
       50350 Mb (max)   | 28652 Mb (max)     1.76x   2.11x
       41159 Mb (rest)  | 19333 Mb (rest)    2.13x   3.15x
        3496 Mb (deb)   |  3496 Mb (deb)      -
        9155 Mb (wic)   |  9161 Mb (wic)      -
       10762 Mb (cache) |   243 Mb (cache)  44.29x
       time: 22m47s     | time: 11m44s       1.94x

The build time improved even more than a real machine with 8 threads CPU and
8GB of real RAM.

HOW TO TEST THE PERFORMANCES
============================

NOTE: these repositories are public PoC and are not suitable for business as-is

    git clone https://github.com/robang74/isar-nvidia-debian

    cd isar-nvidia-debian

    git switch devel        # optional, it should be the default branch

    ./build.sh isar         #
    cd isar                 #
    git switch evo2         # optional, it should be the default branch
    cd ..                   #

    ./build.sh complete

This command will take some time to download all the components and the .deb
packages that are needed. It also requires some system / container components
if the machine was not prepared for this kind of building.

    ./check-build-perform.sh basic-os

    ./check-build-perform.sh complete

These two commands deliver the results of the performances achieved.

To compare with the original version modify the kas.yml but a quick way is:

    cd isar
    git switch next
    touch KEEP.THIS
    cd ..

The 'next' is as nearer to ilbers:next as possible but it works immediately.

WARNING: the first build downloads software from external repositories outside
         my control and - rarely, sometimes they change in such a way they
         break the building. If it happens, do not hesitate to contact me.

Signed-off-by: Roberto A. Foglietta <roberto.foglietta@gmail.com>