[RAF] perf. suite script and presentation of the results v9
THe performance suite scripts and presentation of the results

v.2: some blanks spaces removed in doc/build-optim.md

v.3: new script: check-build-perform.sh [build_target]

v.4: new script: isar-next-rebase.sh [-r]

v.5: isar-next-rebase.sh rpull the local branch before rebasing

v.6: script update: check-build-perform.sh

v.7: performances showcase added in the commit description

v.8: isar-next-rebase.sh updated

v.9: added the benchmarks for cached rebuild + build-optim.md updated

IMAGE CACHE PERFORMANCES
========================

The test suite has been modified because now there are two cases for each
configuration: fresh build and cached rebuild.

        devel2 (public) + npriv5 (private)
       ------------------------------------
            fresh              cache
       ------------ complete -------------- ==============
       32938 Mb (max)   | 28217 Mb (max)     1.17x   1.25x
       23707 Mb (rest)  | 18990 Mb (rest)    1.25x   1.48x
        3537 Mb (deb)   |  3537 Mb (deb)      -
        9143 Mb (wic)   |  9139 Mb (wic)      -
        4650 Mb (cache) |  4650 Mb (cache)    -
       time: 15m15s     | time:  4m37s       3.30x
       ------------------------------------
                    evo2, time: 10m9s        1.50x

The main reason to use image cache is to speed-up the rebuilds but the ratio
depends on the ability of ISAR to deal with files, in fact, using 'evo2' the
public branch parent of 'npriv5' the gain using the cache is quite limited.

Instead with npriv5, the improvement in using the image cache makes it suitable
also for being used in large projects. Further improvements are still possible.

         devel2 (public) + next (public)
       ------------------------------------
          original    cache    nvpriv5
       ------------ complete -------------- ==============
       39651 Mb (max)   | 28217 Mb (max)     1.41x   1.60x
       30461 Mb (rest)  | 18990 Mb (rest)    1.60x   2.16x
        3537 Mb (deb)   |  3537 Mb (deb)      -
        9155 Mb (wic)   |  9139 Mb (wic)      -
       15283 Mb (cache) |  4650 Mb (cache)    -
       time:  9m31s     | time:  4m37s       2.06x

Without any surprise the npriv5 performs 2x better than the original version
also when the image cache is enabled on a large project.

PERFORMANCES SHOWCASES
======================

Compared to the original the performances increase is impressive, in both cases:

          original          roban74:evo2
       ------------ basic-os -------------- ==============
       43954 Mb (max)   |  3498 Mb (max)    12.57x  16.07x
       26548 Mb (rest)  |  2588 Mb (rest)   10.26x  14.50x
        3741 Mb (deb)   |  3417 Mb (deb)      -
         820 Mb (wic)   |   814 Mb (wic)      -
       11789 Mb (cache) |   273 Mb (cache)  43.18x
       time: 8m33s      | time: 3m09s        2.71x

          original          roban74:evo2
       ------------ complete ------------- ===============
       52507 Mb (max)   | 28606 Mb (max)     1.84x   2.23x
       43311 Mb (rest)  | 19413 Mb (rest)    2.23x   3.33x
        3741 Mb (deb)   |  3417 Mb (deb)      -
        9159 Mb (wic)   |  9155 Mb (wic)      -
       11799 Mb (cache) |   283 Mb (cache)  41.69x
       time: 20m13s     | time: 11m44s       1.72x

Considering that the local .deb cache growing with the time because new updates
will be added, and comparing the performance gains between the basic-os and the
complete image, we can say that on the long run also the complete case will
reach the 2x of performance at least in time building.

HOW TO TEST THE PERFORMANCES
============================

NOTE: these repositories are public PoC and are not suitable for business as-is

    git clone https://github.com/robang74/isar-nvidia-debian

    cd isar-nvidia-debian

    git switch devel2       # necessary, default branch is devel

    ./build.sh isar         #
    cd isar                 #
    git switch evo2         # optional, it should be the default branch
    cd ..                   #

    ./build.sh complete

This command will take some time to download all the components and the .deb
packages that are needed. It also requires some system / container components
if the machine was not prepared for this kind of building.

    ./check-build-perform.sh basic-os

    ./check-build-perform.sh complete

These two commands deliver the results of the performances achieved. Each of
them will do 2 tests: a fresh one and the second one leveraging the cache.

To compare with the original version modify the kas.yml but a quick way is:

    cd isar
    git switch next
    touch KEEP.THIS
    cd ..

The 'next' is as nearer to ilbers:next as possible but it works immediately.

WARNING: the first build downloads software from external repositories outside
         my control and - rarely, sometimes they change in such a way they
         break the building. If it happens, do not hesitate to contact me.

Signed-off-by: Roberto A. Foglietta <roberto.foglietta@gmail.com>