dpkg-base: set default for RPROVIDES
This is a follow-up to 40a53789d46763441715508b82154c976700075b

When RDEPENDing on something that is not the PN of a recipe, bitbake needs
a corresponding RPROVIDES.
While Yocto is very precise with definitions of DEPENDS vs. RDEPENDS (and thus
also the RPROVIDES are properly maintained), the Isar way is to specify build
dependencies with DEPENDS, and runtime dependencies with DEBIAN_DEPENDS.
This sets a sane default for RPROVIDES, and most users should be able to
use RDEPENDS without explicitly overriding it.

Acked-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Adriaan Schmidt <adriaan.schmidt@siemens.com>