rootfs_import_package_cache() before rootfs_install_pkgs_update()
At this time the functions dl_deb_im/export() transfer also the debian
repositories sources lists in such a way 'apt-get update' has the
minimum network effort to accomplish or even none at all. However,
because of this, the import of the cache should precede the update.

It depends on this patch or one of its kind/version:

* deb_dl_dir_im/export rework, also transfer the repositories lists v4

Signed-off-by: Roberto A. Foglietta <roberto.foglietta@gmail.com>