sstate_create_package() pipes pigz with -1 -b2048 to be faster
Using pgiz in pipe with tar allows to pass every options allowed
included the -b 2048 that changes the block size from 128k which
is the default in 2M and this improves performances with a very
tiny change in compressed archive size. Larger buffers does not
improve further the performances and smaller starts to be slower
below 1024k.

Signed-off-by: Roberto A. Foglietta <roberto.foglietta@gmail.com>