SSTATE_ZSTD_CLEVEL 2 because a cache should be fast not 12% smaller
sstate.bbclass, improvment

The default set of zstd compression level is changed from 8 to 2 because
the main reason to have a cache is that it should save time, also when
it is created and expecially for large images caching. Set to 2 and not
to 1 because also writing a disk takes time. The difference between 1
and 2 in terms of time is small but enough big in terms of disk usage
to prefer 2 to 1. For the same reasons for which has been choose 8
instead of 9. The extreme values are unbalanced in size/time.

Signed-off-by: Roberto A. Foglietta <roberto.foglietta@gmail.com>