deb-dl-dir, dpkg: every find that hits archives/partial runs with sudo
deb-dl-dir and dpkg, improvment - faster

Now, every find that hits archives/partial runs with sudo and this makes
a huge difference in in term of speed, apparently in an unesplicable
way. However this happens on the top of many changes about the sstate
cache management which in overall reduced the speed but with this change
the complete cached build reached the impressive 3m47s time when at its
best it was 4m11s.

TODO: brings back to an older version and check again
DONE: works faster after the cache management rewrite

Signed-off-by: Roberto A. Foglietta <roberto.foglietta@gmail.com>