Reinstall imager deps if package in IMAGER_INSTALL has changed.
When debugging packages in `isar-apt` that are in IMAGER_INSTALL
list, they happen not to be updated in the resulting image.

For example, in case of u-boot IMAGE_BOOT_FILES is used to generate
files for `/boot` partition. Buildchroot is currently used for
extracting required files from `u-boot`, `optee-os`, etc. These
packages are installed in buildchroot via IMAGER_INSTALL list.

Since modifications of the packages don't change their versions,
updated packages don't go to buildchroot, so previously used versions
are wrongly used. The following can be seen in log:

"<package> is already the newest version"

This patch forces reinstalling all the packages from IMAGER_INSTALL
list if some of them has changed.

Signed-off-by: Uladzimir Bely <ubely@ilbers.de>